Fix Python dependency updating bug
The replacement of packages in Python requirements file would replace
packages that contained the names of other packages with the shorter
package name, because it wasn't matching the end of the name. Example:

django-allauth==0.27.0
django-allauth-2fa==0.2

Would be replaced with:

django-allauth==0.28.0
django-allauth==0.28.0

The same logic in a different place would cause a similar error with a
"skip" comment for a package with a long name skipping updating packages
with shorter names that are substrings of the longer package name.

Change searches to require "==" immediately after the end of the package
name.